+++ src/main/java/com/pinkproject/admin/AdminController.java
@@ -66,205 +66,4 @@ public class AdminController {
         return "redirect:/admin";
     }
 
-    @GetMapping("/admin/notice")
-    public String notice(@RequestParam(value = "keyword", required = false) String keyword,
-                         @RequestParam(value = "page", defaultValue = "1") int page,
-                         HttpServletRequest request) {
-        int pageIndex = page - 1; // 페이지 인덱스를 0부터 시작하도록 변환
-        Page<_DetailNoticeAdminRecord> notices = keyword != null && !keyword.isEmpty() ?
-                noticeService.searchNotices(keyword, pageIndex) :
-                noticeService.getNotices(pageIndex);
-        request.setAttribute("notices", notices.getContent());
-
-        // 페이지네이션 정보 추가
-        request.setAttribute("currentPage", page);
-        request.setAttribute("totalPages", notices.getTotalPages());
-        request.setAttribute("hasPrevious", notices.hasPrevious());
-        request.setAttribute("hasNext", notices.hasNext());
-        request.setAttribute("previousPage", page > 1 ? page - 1 : 1); // 1 페이지 이하로 내려가지 않도록
-        request.setAttribute("nextPage", page + 1);
-        request.setAttribute("lastPage", notices.getTotalPages());
-        request.setAttribute("pages", IntStream.range(1, notices.getTotalPages() + 1)
-                .mapToObj(i -> Map.of("number", i, "isCurrent", i == page))
-                .collect(Collectors.toList()));
-        request.setAttribute("keyword", keyword);
-
-        // 세션에서 admin 객체 가져와서 username 설정
-        SessionAdmin sessionAdmin = (SessionAdmin) session.getAttribute("admin");
-        if (sessionAdmin != null) {
-            request.setAttribute("username", sessionAdmin.getUsername());
-        } else {
-            request.setAttribute("username", ""); // 기본값 설정
-        }
-
-        // currentDateTime 설정
-        LocalDateTime now = LocalDateTime.now();
-        DateTimeFormatter formatter = DateTimeFormatter.ofPattern("yyyy-MM-dd");
-        String formattedNow = now.format(formatter);
-        request.setAttribute("currentDateTime", formattedNow);
-
-        return "admin/notice";
-    }
-
-
-    @GetMapping("/admin/faq")
-    public String faq(@RequestParam(value = "keyword", required = false) String keyword,
-                      @RequestParam(value = "page", defaultValue = "1") int page,
-                      HttpServletRequest request) {
-        int pageIndex = page - 1; // 페이지 인덱스를 0부터 시작하도록 변환
-        Page<_DetailFaqAdminRecord> faqs = keyword != null && !keyword.isEmpty() ?
-                faqService.searchFaqs(keyword, pageIndex) :
-                faqService.getFaqs(pageIndex);
-        request.setAttribute("faqs", faqs.getContent());
-
-        // 페이지네이션 정보 추가
-        request.setAttribute("currentPage", page);
-        request.setAttribute("totalPages", faqs.getTotalPages());
-        request.setAttribute("hasPrevious", faqs.hasPrevious());
-        request.setAttribute("hasNext", faqs.hasNext());
-        request.setAttribute("previousPage", page > 1 ? page - 1 : 1); // 1 페이지 이하로 내려가지 않도록
-        request.setAttribute("nextPage", page + 1);
-        request.setAttribute("lastPage", faqs.getTotalPages());
-        request.setAttribute("pages", IntStream.range(1, faqs.getTotalPages() + 1) // 1부터 시작하도록 수정
-                .mapToObj(i -> Map.of("number", i, "isCurrent", i == page))
-                .collect(Collectors.toList()));
-        request.setAttribute("keyword", keyword);
-
-        // 세션에서 admin 객체 가져와서 username 설정
-        SessionAdmin sessionAdmin = (SessionAdmin) session.getAttribute("admin");
-        if (sessionAdmin != null) {
-            request.setAttribute("username", sessionAdmin.getUsername());
-        } else {
-            request.setAttribute("username", ""); // 기본값 설정
-        }
-
-        // currentDateTime 설정
-        LocalDateTime now = LocalDateTime.now();
-        DateTimeFormatter formatter = DateTimeFormatter.ofPattern("yyyy-MM-dd");
-        String formattedNow = now.format(formatter);
-        request.setAttribute("currentDateTime", formattedNow);
-
-        return "admin/faq";
-    }
-
-
-
-    @GetMapping("/admin/faq/detail/{id}")
-    public String faqDetail(@PathVariable Integer id, HttpServletRequest request) {
-        _DetailFaqAdminRecord faqDetail = faqService.detailFaqAdminRecord(id);
-        HttpSession session = request.getSession();
-        SessionAdmin sessionAdmin = (SessionAdmin) session.getAttribute("admin");
-        if (sessionAdmin != null) {
-            request.setAttribute("username", sessionAdmin.getUsername());
-            LocalDateTime now = LocalDateTime.now();
-            DateTimeFormatter formatter = DateTimeFormatter.ofPattern("yyyy-MM-dd");
-            String formattedNow = now.format(formatter);
-            request.setAttribute("currentDateTime", formattedNow);
-        }
-        request.setAttribute("faq", faqDetail);
-        request.setAttribute("title", faqDetail.title());
-        request.setAttribute("content", faqDetail.content());
-        request.setAttribute("date", faqDetail.date());
-
-        LocalDateTime now = LocalDateTime.now();
-        DateTimeFormatter formatter = DateTimeFormatter.ofPattern("yyyy-MM-dd");
-        String formattedNow = now.format(formatter);
-        request.setAttribute("currentDateTime", formattedNow);
-
-        return "admin/faq-detail";
-    }
-
-    @GetMapping("/admin/notice/detail/{id}")
-    public String noticeDetail(@PathVariable Integer id, HttpServletRequest request) {
-        _DetailNoticeAdminRecord noticeDetail = noticeService.detailNoticeAdminRecord(id);
-        HttpSession session = request.getSession(); // 세션 객체를 얻음
-        SessionAdmin sessionAdmin = (SessionAdmin) session.getAttribute("admin");
-        if (sessionAdmin != null) {
-            request.setAttribute("username", sessionAdmin.getUsername());
-            LocalDateTime now = LocalDateTime.now();
-            DateTimeFormatter formatter = DateTimeFormatter.ofPattern("yyyy-MM-dd");
-            String formattedNow = now.format(formatter);
-            request.setAttribute("currentDateTime", formattedNow);
-        }
-        request.setAttribute("notice", noticeDetail);
-        request.setAttribute("title", noticeDetail.title());
-        request.setAttribute("content", noticeDetail.content());
-        request.setAttribute("date", noticeDetail.date());
-        return "admin/notice-detail";
-    }
-
-    @DeleteMapping("/admin/faq/delete/{id}")
-    @ResponseBody
-    public ResponseEntity<?> deleteFaq(@PathVariable Integer id) {
-        try {
-            faqService.deleteFaq(id);
-            return ResponseEntity.ok(new ApiUtil<>(null));
-        } catch (Exception e) {
-            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(new ApiUtil<>(HttpStatus.INTERNAL_SERVER_ERROR.value(), e.getMessage()));
-        }
-    }
-
-    @DeleteMapping("/admin/notice/delete/{id}")
-    @ResponseBody
-    public ResponseEntity<?> deleteNotice(@PathVariable Integer id) {
-        try {
-            noticeService.deleteNotice(id);
-            return ResponseEntity.ok(new ApiUtil<>(null));
-        } catch (Exception e) {
-            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(new ApiUtil<>(HttpStatus.INTERNAL_SERVER_ERROR.value(), e.getMessage()));
-        }
-    }
-
-
-    @GetMapping("/admin/faq/save")
-    public String saveFaqForm(HttpServletRequest request) {
-        SessionAdmin sessionAdmin = (SessionAdmin) session.getAttribute("admin");
-        if (sessionAdmin != null) {
-            request.setAttribute("username", sessionAdmin.getUsername());
-            LocalDateTime now = LocalDateTime.now();
-            DateTimeFormatter formatter = DateTimeFormatter.ofPattern("yyyy-MM-dd");
-            String formattedNow = now.format(formatter);
-            request.setAttribute("currentDateTime", formattedNow);
-        }
-        return "admin/faq-save";
-    }
-
-    @PostMapping("/admin/faq/save")
-    public String saveFaq(@ModelAttribute _SaveFaqAdminRecord saveFaqAdminRecord, HttpServletRequest request) {
-        try {
-            SessionAdmin sessionAdmin = (SessionAdmin) session.getAttribute("admin");
-            Admin admin = adminService.findByUsername(sessionAdmin.getUsername());
-            faqService.saveFaq(saveFaqAdminRecord, admin);
-            return "redirect:/admin/faq";
-        } catch (Exception e) {
-            request.setAttribute("errorMessage", e.getMessage());
-            return "admin/faq-save";
-        }
-    }
-    @GetMapping("/admin/notice/save")
-    public String saveNoticeForm(HttpServletRequest request) {
-        SessionAdmin sessionAdmin = (SessionAdmin) session.getAttribute("admin");
-        if (sessionAdmin != null) {
-            request.setAttribute("username", sessionAdmin.getUsername());
-            LocalDateTime now = LocalDateTime.now();
-            DateTimeFormatter formatter = DateTimeFormatter.ofPattern("yyyy-MM-dd");
-            String formattedNow = now.format(formatter);
-            request.setAttribute("currentDateTime", formattedNow);
-        }
-        return "admin/notice-save";
-    }
-
-    @PostMapping("/admin/notice/save")
-    public String saveNotice(@ModelAttribute _SaveNoticeAdminRecord saveNoticeAdminRecord, HttpServletRequest request) {
-        try {
-            SessionAdmin sessionAdmin = (SessionAdmin) session.getAttribute("admin");
-            Admin admin = adminService.findByUsername(sessionAdmin.getUsername());
-            noticeService.saveNotice(saveNoticeAdminRecord, admin);
-            return "redirect:/admin/notice";
-        } catch (Exception e) {
-            request.setAttribute("errorMessage", e.getMessage());
-            return "admin/notice-save";
-        }
-    }
-
 }
