+++ src/test/java/com/pinkproject/admin/AdminServiceTest.java
@@ -1,0 +1,4 @@
+package com.pinkproject.admin;
+
+public class AdminServiceTest {
+}
